Lifted out the users table custom component to be reusable
